Rename blocks to be shorter
