+++ packages/builder/src/components/design/AppPreview/componentStructure.json
@@ -3,8 +3,8 @@
     "name": "Blocks",
     "icon": "Article",
     "children": [
-      "tablewithsearch",
-      "cardlistwithsearch"
+      "tableblock",
+      "cardsblock"
     ]
   },
   "section",
