+++ packages/client/src/components/app/blocks/CardsBlock.svelte
