+++ packages/client/src/components/app/blocks/TableBlock.svelte
